bring over some files still in use
