+++ shared/profile/search.js.flow
@@ -1,7 +1,7 @@
 // @flow
 import {Component} from 'react'
 
-import type {SearchResultId} from '../constants/searchv3'
+import type {SearchResultId} from '../constants/search'
 import type {UserDetails} from '../search/user-input'
 
 export type Props = {
